+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased][unreleased]
 
+- Fix nullable field long-form
+
 ## [1.3.4][] - 2021-09-10
 
 - Show path to the field in warnings